Update previously generated directories gracefully.
Generated files are logged in a ".webgenerator.manifest" file.
When that file is presented in a destination dir, all files mentioned in
it are deleted. Then, the new website is merged into the destination dir.
That way, files and directories that should not be touched will stay intact,
like a .git dir, or in our case the CNAME file OR if anyone added files
manually.
NOTE: The generator will override files that it generated even if they are
not mentioned in the manifest file..